service/requirements.txt
google-cloud-storage
Flask
gunicorn
urllib3
requests